fix(edid/color): flipped two smallest bits
these were resulting in slight differences in the decimal values.

after this change, i can now use a tolerance of 0.0005 (as suggested by the standard) to test the decimal values

+++ edid/src/parser/basic_info.rs
@@ -201,7 +201,7 @@ fn gamma(input: &[u8]) -> Option<Decimal> {
         }
 
         // reverse from the standard: byte = (GAMMA x 100) – 100
-        Some((Decimal::from(byte) + 100) / 100)
+        Some((Decimal::from(byte) + 100) / Decimal::from(100))
     }
 }
